Fix typing, inheritance and static checks
